+++ src/controllers/SettingsController.php
@@ -59,9 +59,9 @@ class SettingsController extends Controller
                     'error' => $errorMessage
                 ]);
             } else {
-                Craft::$app->user->setError($errorMessage);
+                Craft::$app->getSession()->setError($errorMessage);
 
-                Craft::$app->urlManager->setRouteVariables([
+                Craft::$app->getUrlManager()->setRouteParams([
                     'errorCode' => $errorCode,
                     'errorMessage' => $errorMessage,
                 ]);